fixed sdk tests
