+++ sdk/src/test/java/io/customer/sdk/data/store/DeviceStoreTest.kt
@@ -1,13 +1,10 @@
 package io.customer.sdk.data.store
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import io.customer.commontest.BaseTest
+import io.customer.commontest.BaseUnitTest
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@RunWith(AndroidJUnit4::class)
-class DeviceStoreTest : BaseTest() {
+class DeviceStoreTest : BaseUnitTest() {
 
     @Test
     fun verifyBuildAttributesInDevice() {
