+++ sdk/src/test/java/io/customer/sdk/queue/QueueTest.kt
@@ -1,8 +1,7 @@
 package io.customer.sdk.queue
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.squareup.moshi.JsonClass
-import io.customer.commontest.BaseTest
+import io.customer.commontest.BaseUnitTest
 import io.customer.sdk.extensions.random
 import io.customer.sdk.queue.type.QueueModifyResult
 import io.customer.sdk.queue.type.QueueStatus
@@ -15,12 +14,17 @@ import org.amshove.kluent.shouldBeEqualTo
 import org.amshove.kluent.shouldBeNull
 import org.junit.Before
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.mockito.kotlin.*
+import org.mockito.kotlin.any
+import org.mockito.kotlin.anyOrNull
+import org.mockito.kotlin.eq
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.never
+import org.mockito.kotlin.nullableArgumentCaptor
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.whenever
 
 @ExperimentalCoroutinesApi
-@RunWith(AndroidJUnit4::class)
-class QueueTest : BaseTest() {
+class QueueTest : BaseUnitTest() {
 
     private lateinit var queue: QueueImpl
     val storageMock: QueueStorage = mock()
