+++ sdk/src/test/java/io/customer/sdk/util/JsonAdapterTest.kt
@@ -1,8 +1,7 @@
 package io.customer.sdk.util
 
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.squareup.moshi.JsonClass
-import io.customer.commontest.BaseTest
+import io.customer.commontest.BaseUnitTest
 import org.amshove.kluent.AnyException
 import org.amshove.kluent.invoking
 import org.amshove.kluent.shouldBeEqualTo
@@ -10,10 +9,8 @@ import org.amshove.kluent.shouldBeNull
 import org.amshove.kluent.shouldNotBeNull
 import org.amshove.kluent.shouldThrow
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@RunWith(AndroidJUnit4::class)
-class JsonAdapterTest : BaseTest() {
+class JsonAdapterTest : BaseUnitTest() {
 
     @JsonClass(generateAdapter = true)
     data class TestVo(
